Re-use test utilities
